管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -33,6 +33,7 @@ public class EmployeeAction extends ActionBase {
      */
     public void index() throws ServletException, IOException {
 
+        if(checkAdmin()) {
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
 
@@ -50,26 +51,19 @@ public class EmployeeAction extends ActionBase {
         }
 
         forward(ForwardConst.FW_EMP_INDEX);
-
+        }
     }
-    /**
-     * @throws ServletException
-     * @throws IOException
-     */
     public void entryNew() throws ServletException, IOException {
 
+        if(checkAdmin()) {
         putRequestScope(AttributeConst.TOKEN, getTokenId());
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
 
         forward(ForwardConst.FW_EMP_NEW);
-    }
-    /**
-     * @throws ServletException
-     * @throws IOException
-     */
+    }}
     public void create() throws ServletException, IOException {
 
-        if (checkToken()) {
+        if (checkToken() && checkToken()) {
 
             EmployeeView ev = new EmployeeView(
                     null,
@@ -108,6 +102,7 @@ public class EmployeeAction extends ActionBase {
      */
     public void show() throws ServletException, IOException {
 
+        if (checkAdmin()) {
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
         if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
@@ -119,13 +114,14 @@ public class EmployeeAction extends ActionBase {
         putRequestScope(AttributeConst.EMPLOYEE, ev);
 
         forward(ForwardConst.FW_EMP_SHOW);
-    }
+    }}
     /**
      * @throws ServletException
      * @throws IOException
      */
     public void edit() throws ServletException, IOException {
 
+        if (checkAdmin()) {
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
         if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
@@ -138,7 +134,7 @@ public class EmployeeAction extends ActionBase {
         putRequestScope(AttributeConst.EMPLOYEE, ev);
 
         forward(ForwardConst.FW_EMP_EDIT);
-
+        }
     }
     /**
      * @throws ServletException
@@ -146,7 +142,7 @@ public class EmployeeAction extends ActionBase {
      */
     public void update() throws ServletException, IOException {
 
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
                     getRequestParam(AttributeConst.EMP_CODE),
@@ -176,13 +172,9 @@ public class EmployeeAction extends ActionBase {
             }
         }
     }
-    /**
-     * @throws ServletException
-     * @throws IOException
-     */
     public void destroy() throws ServletException, IOException {
 
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
 
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -190,6 +182,21 @@ public class EmployeeAction extends ActionBase {
 
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
         }
+    }
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+
+        } else {
+
+            return true;
+        }
+
     }
 
 }
